server/scanners/urls.py
# scanners/urls.py

from django.urls import path, include
from rest_framework.routers import DefaultRouter

from .views import ScannerViewSet

router = DefaultRouter()
router.register(
    prefix   = "",
    viewset  = ScannerViewSet,
    basename = "scanners",
)

urlpatterns = [
    path("", include(router.urls)),
]