server/requirements.txt
amqp==5.3.1
asgiref==3.11.1
attrs==26.1.0
autobahn==25.12.2
Automat==25.4.16
billiard==4.2.4
black==26.3.1
cbor2==5.8.0
celery==5.6.2
certifi==2026.2.25
cffi==2.0.0
channels==4.3.2
channels_redis==4.3.0
charset-normalizer==3.4.6
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
cloudinary==1.44.1
colorama==0.4.6
constantly==23.10.4
cron-descriptor==1.4.5
cryptography==46.0.5
daphne==4.2.1
Django==6.0.3
django-anymail==14.0
django-celery-beat==2.9.0
django-cloudinary-storage==0.3.0
django-cors-headers==4.9.0
django-environ==0.13.0
django-extensions==4.1
django-filter==25.2
django-ratelimit==4.1.0
django-timezone-field==7.2.1
django_celery_results==2.6.0
djangorestframework==3.17.0
djangorestframework_simplejwt==5.5.1
drf-spectacular==0.29.0
factory_boy==3.3.3
Faker==40.11.0
flake8==7.3.0
hyperlink==21.0.0
idna==3.11
Incremental==24.11.0
inflection==0.5.1
iniconfig==2.3.0
isort==8.0.1
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
kombu==5.6.2
mccabe==0.7.0
msgpack==1.1.2
mypy_extensions==1.1.0
packaging==26.0
pathspec==1.0.4
phonenumbers==9.0.26
pillow==12.1.1
platformdirs==4.9.4
pluggy==1.6.0
prompt_toolkit==3.0.52
psycopg2-binary==2.9.11
py-ubjson==0.16.1
pyasn1==0.6.3
pyasn1_modules==0.4.2
pycodestyle==2.14.0
pycparser==3.0
pyflakes==3.4.0
Pygments==2.19.2
PyJWT==2.12.1
pyOpenSSL==26.0.0
pytest==9.0.2
pytest-django==4.12.0
python-barcode==0.16.1
python-crontab==3.3.0
python-dateutil==2.9.0.post0
python-slugify==8.0.4
pytokens==0.4.1
PyYAML==6.0.3
redis==7.3.0
referencing==0.37.0
reportlab==4.4.10
requests==2.32.5
rpds-py==0.30.0
sentry-sdk==2.55.0
service-identity==24.2.0
six==1.17.0
sqlparse==0.5.5
stripe==14.4.1
text-unidecode==1.3
Twisted==25.5.0
txaio==25.12.2
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
ujson==5.12.0
uritemplate==4.2.0
urllib3==2.6.3
vine==5.1.0
wcwidth==0.6.0
zope.interface==8.2
